Remove get_errno.c from the funit-core sources.
If a C source file is present in the list of source files the VS project will be a C project rather than a Fortran. As a result the Fortran source files will not be compiled. It seems that get_errno.c is not used anywhere so can be safely omitted.