Fix slow try_compile() and try_run() in CMake
try_compile() and try_run() are used to check the compiler capabilities and internal representation of certain data types. However, both are _extremely_ slow with the VS generator; see issue here: https://discourse.cmake.org/t/cmake-configuration-and-generate-is-extremely-slow-for-vs-2017/2036. The problem is that, for every try_compile()/tr_run() call, VS project and solution files with multiple configurations are generated. As a workaround in this commit, compiling and running is "manually" using CMake function execute_process().
Note:
* this probably breaks things on non-windows systems or non-VS build systems
* it is required that the correct environment variables (e.g. INCLUDE, LIB) are set during CMake configuration. This is not (necessarily) the case when running configuration within VS (due to ZERO_CHECK). This means that CMake configuration will have to be run from a command prompt with the correct environment variables set.