fhamcrest generated sources are now created during CMake configuration
Previously sources were generated as part of the build process. This led to problems when building with Intel Visual Fortran because these sources did not exist at the start of the fhamcrest project so dependencies could not be scanned. This caused incorrect compilation order of the sources and module-not-found errors.